+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+- Updated to `0.3.x` for `tracing-subscriber`.
+- Refactored the token registry to fix an issue with `arc-swap` needing to allocate on the read
+  path, which caused reentrancy during allocation tracking.
+
+## [0.1.2] - 2021-10-04
+
 ### Added
 - Ability to specify a custom allocator to wrap around instead of always using the system allocator.